permutation wip

+++ Coursework/Coursework1/main.cpp
@@ -1,13 +1,13 @@
 /** @file */
 
-#include <iostream>
 #include <ctime>
+#include <iostream>
 
 #include "Coursework1Lib/CW1CommandLine.h"
 
 int main()
 {
-  std::srand(std::time(0));
+  std::srand((unsigned int)std::time(0));
   Coursework1::CW1CommandLine cli(std::cin, std::cout);
   cli.initCLI();
   return cli.run();
